fix(protection): cast i64 index to i32 for ONNX tensor creation
- ONNX models actually expect int32, not int64
- Accept i64 parameter for API consistency, cast to i32 for tensor
- Fixes 'Unexpected input data type' error
- Allow Glaze and Nightshade protection to proceed with correct types

+++ src-tauri/src/onnx_integration/protection/algorithms.rs
@@ -65,8 +65,8 @@ fn create_image_tensor(input: &Array4<f32>) -> Result<Tensor<f32>, String> {
     .map_err(|e| format!("Failed to create input tensor: {}", e))
 }
 
-fn create_index_tensor(index: i64) -> Result<Tensor<i64>, String> {
-    Tensor::from_array(([1_usize], vec![index].into_boxed_slice()))
+fn create_index_tensor(index: i64) -> Result<Tensor<i32>, String> {
+    Tensor::from_array(([1_usize], vec![index as i32].into_boxed_slice()))
         .map_err(|e| format!("Failed to create index tensor: {}", e))
 }
